test_trashcan() and supporting class Ouch():  Jeremy noted that this test takes much longer to run in the context of the test suite than when run in isolation.  That's because it forces a large number of full collections, which take time proportional to the total number of gc'ed objects in the whole system.
But since the dangerous implementation trickery that caused this test to
fail in 2.0, 2.1 and 2.2 doesn't exist in 2.3 anymore (the trashcan
mechanism stopped doing evil things when the possibility for compiling
without cyclic gc was taken away), such an expensive test is no longer
justified.  This checkin leaves the test intact, but fiddles the
constants to reduce the runtime by about a factor of 5.

--HG--
branch : legacy-trunk