Adapted for the move of Mac/Lib to Lib/plat-mac. Makefile.pre.in now knows about plat-mac subdirectories, and configure adds a variable EXTRAPLATDIR. These together take care of copying Lib/plat-mac to the destination on darwin.
Adding plat-mac is still done with a .pth file which is only created when
you do a framework build. I'm not 100% happy with this, but fixing it
really needs a functional pythonw in non-framework builds, and I don't
think I can do that before 2.3a1 (but I'll try:-).

--HG--
branch : legacy-trunk